+++ src/attr.rs
@@ -321,7 +321,7 @@ impl Rewrite for ast::Attribute {
         } else {
             let should_skip = self
                 .ident()
-                .map(|s| context.skip_context.attributes_skip(&*s.name.as_str()))
+                .map(|s| context.skip_context.skip_attribute(&s.name.as_str()))
                 .unwrap_or(false);
             let prefix = attr_prefix(self);