+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## [1.3.3] 2019-07-15
 
+### Added
+
+- Add new attribute `rustfmt::skip::attributes` to prevent rustfmt 
+from formatting an attribute #3665
+
 ### Fixed
 
 - Fix `cargo fmt -- --help` printing nothing (#3620).